+++ lib/src/visitor/evaluate.dart
@@ -5,13 +5,14 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: fa9a95be772a0bb1e769db23a3d794c6c55148d4
+// Checksum: 81d67b4973f0c87105e0c64d39b540dd82af1549
 //
 // ignore_for_file: unused_import
 
 import 'async_evaluate.dart' show EvaluateResult;
 export 'async_evaluate.dart' show EvaluateResult;
 
+import 'dart:collection';
 import 'dart:math' as math;
 
 import 'package:charcode/charcode.dart';
@@ -314,6 +315,13 @@ class _EvaluateVisitor
   /// stylesheet.
   List<ModifiableCssImport>? _outOfOrderImports;
 
+  /// A map from modules loaded by the current module to loud comments written
+  /// in this module that should appear before the loaded module.
+  ///
+  /// This is `null` unless there are any pre-module comments in the current
+  /// stylesheet.
+  Map<Module<Callable>, List<CssComment>>? _preModuleComments;
+
   /// The extension store that tracks extensions and style rules for the current
   /// module.
   ExtensionStore get _extensionStore =>
@@ -503,7 +511,7 @@ class _EvaluateVisitor
         }
 
         _loadModule(url, "load-css()", callableNode,
-            (module) => _combineCss(module, clone: true).accept(this),
+            (module, _) => _combineCss(module, clone: true).accept(this),
             baseUrl: callableNode.span.sourceUrl,
             configuration: configuration,
             namesInErrors: true);
@@ -583,7 +591,9 @@ class _EvaluateVisitor
     }
   }
 
-  /// Loads the module at [url] and passes it to [callback].
+  /// Loads the module at [url] and passes it to [callback], along with a
+  /// boolean indicating whether this is the first time the module has been
+  /// loaded.
   ///
   /// This first tries loading [url] relative to [baseUrl], which defaults to
   /// `_stylesheet.span.sourceUrl`.
@@ -600,7 +610,7 @@ class _EvaluateVisitor
   /// The [stackFrame] and [nodeWithSpan] are used for the name and location of
   /// the stack frame for the duration of the [callback].
   void _loadModule(Uri url, String stackFrame, AstNode nodeWithSpan,
-      void callback(Module<Callable> module),
+      void callback(Module<Callable> module, bool firstLoad),
       {Uri? baseUrl,
       Configuration? configuration,
       bool namesInErrors = false}) {
@@ -614,7 +624,9 @@ class _EvaluateVisitor
             configuration.nodeWithSpan.span);
       }
 
-      _addExceptionSpan(nodeWithSpan, () => callback(builtInModule));
+      // Always consider built-in stylesheets to be "already loaded", since they
+      // never require additional execution to load and never produce CSS.
+      _addExceptionSpan(nodeWithSpan, () => callback(builtInModule, false));
       return;
     }
 
@@ -637,6 +649,7 @@ class _EvaluateVisitor
       }
       if (canonicalUrl != null) _activeModules[canonicalUrl] = nodeWithSpan;
 
+      var firstLoad = !_modules.containsKey(canonicalUrl);
       var oldInDependency = _inDependency;
       _inDependency = result.isDependency;
       Module<Callable> module;
@@ -650,7 +663,7 @@ class _EvaluateVisitor
         _inDependency = oldInDependency;
       }
 
-      _addExceptionSpan(nodeWithSpan, () => callback(module),
+      _addExceptionSpan(nodeWithSpan, () => callback(module, firstLoad),
           addStackFrame: false);
     });
   }
@@ -699,11 +712,13 @@ class _EvaluateVisitor
 
     var environment = Environment();
     late CssStylesheet css;
+    late Map<Module<Callable>, List<CssComment>>? preModuleComments;
     var extensionStore = ExtensionStore();
     _withEnvironment(environment, () {
       var oldImporter = _importer;
       var oldStylesheet = __stylesheet;
       var oldRoot = __root;
+      var oldPreModuleComments = _preModuleComments;
       var oldParent = __parent;
       var oldEndOfImports = __endOfImports;
       var oldOutOfOrderImports = _outOfOrderImports;
@@ -734,10 +749,12 @@ class _EvaluateVisitor
       css = _outOfOrderImports == null
           ? root
           : CssStylesheet(_addOutOfOrderImports(), stylesheet.span);
+      preModuleComments = _preModuleComments;
 
       _importer = oldImporter;
       __stylesheet = oldStylesheet;
       __root = oldRoot;
+      _preModuleComments = oldPreModuleComments;
       __parent = oldParent;
       __endOfImports = oldEndOfImports;
       _outOfOrderImports = oldOutOfOrderImports;
@@ -751,7 +768,8 @@ class _EvaluateVisitor
       _configuration = oldConfiguration;
     });
 
-    var module = environment.toModule(css, extensionStore);
+    var module = environment.toModule(
+        css, preModuleComments ?? const {}, extensionStore);
     if (url != null) {
       _modules[url] = module;
       _moduleConfigurations[url] = _configuration;
@@ -793,12 +811,6 @@ class _EvaluateVisitor
       return root.css;
     }
 
-    var sortedModules = _topologicalModules(root);
-    if (clone) {
-      sortedModules = sortedModules.map((module) => module.cloneCss()).toList();
-    }
-    _extendModules(sortedModules);
-
     // The imports (and comments between them) that should be included at the
     // beginning of the final document.
     var imports = <CssNode>[];
@@ -806,19 +818,47 @@ class _EvaluateVisitor
     // The CSS statements in the final document.
     var css = <CssNode>[];
 
-    for (var module in sortedModules.reversed) {
+    /// The modules in reverse topological order.
+    var sorted = Queue<Module<Callable>>();
+
+    /// The modules that have been visited so far. Note that if [cloneCss] is
+    /// true, this contains the original modules, not the copies.
+    var seen = <Module<Callable>>{};
+
+    void visitModule(Module<Callable> module) {
+      if (!seen.add(module)) return;
+      if (clone) module = module.cloneCss();
+
+      for (var upstream in module.upstream) {
+        if (upstream.transitivelyContainsCss) {
+          var comments = module.preModuleComments[upstream];
+          if (comments != null) {
+            // Intermix the top-level comments with plain CSS `@import`s until we
+            // start to have actual CSS defined, at which point start treating it as
+            // normal CSS.
+            (css.isEmpty ? imports : css).addAll(comments);
+          }
+          visitModule(upstream);
+        }
+      }
+
+      sorted.addFirst(module);
       var statements = module.css.children;
       var index = _indexAfterImports(statements);
       imports.addAll(statements.getRange(0, index));
       css.addAll(statements.getRange(index, statements.length));
     }
 
+    visitModule(root);
+
+    if (root.transitivelyContainsExtensions) _extendModules(sorted);
+
     return CssStylesheet(imports + css, root.css.span);
   }
 
-  /// Extends the selectors in each module with the extensions defined in
-  /// downstream modules.
-  void _extendModules(List<Module<Callable>> sortedModules) {
+  /// Destructively updates the selectors in each module with the extensions
+  /// defined in downstream modules.
+  void _extendModules(Iterable<Module<Callable>> sortedModules) {
     // All the [ExtensionStore]s directly downstream of a given module (indexed
     // by its canonical URL). It's important that we create this in topological
     // order, so that by the time we're processing a module we've already filled
@@ -874,33 +914,6 @@ class _EvaluateVisitor
         'The target selector was not found.\n'
         'Use "@extend ${extension.target} !optional" to avoid this error.',
         extension.span);
-  }
-
-  /// Returns all modules transitively used by [root] in topological order,
-  /// ignoring modules that contain no CSS.
-  List<Module<Callable>> _topologicalModules(Module<Callable> root) {
-    // Construct a topological ordering using depth-first traversal, as in
-    // https://en.wikipedia.org/wiki/Topological_sorting#Depth-first_search.
-    var seen = <Module<Callable>>{};
-    var sorted = QueueList<Module<Callable>>();
-
-    void visitModule(Module<Callable> module) {
-      // Each module is added to the beginning of [sorted], which means the
-      // returned list contains sibling modules in the opposite order of how
-      // they appear in the document. Then when the list is reversed to generate
-      // the CSS, they're put back in their original order.
-      for (var upstream in module.upstream) {
-        if (upstream.transitivelyContainsCss && seen.add(upstream)) {
-          visitModule(upstream);
-        }
-      }
-
-      sorted.addFirst(module);
-    }
-
-    visitModule(root);
-
-    return sorted;
   }
 
   /// Returns the index of the first node in [statements] that comes after all
@@ -1348,7 +1361,8 @@ class _EvaluateVisitor
       var newConfiguration =
           _addForwardConfiguration(adjustedConfiguration, node);
 
-      _loadModule(node.url, "@forward", node, (module) {
+      _loadModule(node.url, "@forward", node, (module, firstLoad) {
+        if (firstLoad) _registerCommentsForModule(module);
         _environment.forwardModule(module, node);
       }, configuration: newConfiguration);
 
@@ -1372,7 +1386,8 @@ class _EvaluateVisitor
       _assertConfigurationIsEmpty(newConfiguration);
     } else {
       _configuration = adjustedConfiguration;
-      _loadModule(node.url, "@forward", node, (module) {
+      _loadModule(node.url, "@forward", node, (module, firstLoad) {
+        if (firstLoad) _registerCommentsForModule(module);
         _environment.forwardModule(module, node);
       });
       _configuration = oldConfiguration;
@@ -1407,6 +1422,17 @@ class _EvaluateVisitor
     } else {
       return Configuration.implicit(newValues);
     }
+  }
+
+  /// Adds any comments in [_root.children] to [_preModuleComments] for
+  /// [module].
+  void _registerCommentsForModule(Module<Callable> module) {
+    if (_root.children.isEmpty || !module.transitivelyContainsCss) return;
+    (_preModuleComments ??= {})
+        .putIfAbsent(module, () => [])
+        .addAll(_root.children.cast<CssComment>());
+    _root.clearChildren();
+    _endOfImports = 0;
   }
 
   /// Remove configured values from [upstream] that have been removed from
@@ -2108,7 +2134,8 @@ class _EvaluateVisitor
       configuration = ExplicitConfiguration(values, node);
     }
 
-    _loadModule(node.url, "@use", node, (module) {
+    _loadModule(node.url, "@use", node, (module, firstLoad) {
+      if (firstLoad) _registerCommentsForModule(module);
       _environment.addModule(module, node, namespace: node.namespace);
     }, configuration: configuration);
     _assertConfigurationIsEmpty(configuration);